+++ test/downloader/exit-on-task-worker.js
@@ -1,0 +1,9 @@
+import {parentPort, workerData} from 'node:worker_threads';
+
+const {taskPort} = workerData.workerChannels;
+
+parentPort.postMessage({type: 'ready'});
+
+taskPort.addListener('message', () => {
+  process.exit(1);
+});